pyproject.toml
[project]
name = "gemini-data-extractor"
version = "0.1.0"
description = "A simple tool to extract data from data from PDFs and images using Google's gemini-2.0-flash-001 model."
authors = [
    {name = "Domenico Delle Side", email = "nico@delleside.org"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.3.0",
]

[project.scripts]
gemini-data-extractor = "gemini_data_extractor:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
